backend/requirements.txt
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
tensorflow==1.5.0
keras==2.2
numpy
matplotlib
Pillow
sklearn
theano
scikit-image
pandas
flask
flask_restplus
sqlalchemy
kafka-python
